+++ lib/src/main/java/com/wjduquette/joe/bert/Compiler.java
@@ -5,8 +5,8 @@ import com.wjduquette.joe.SyntaxError;
 import com.wjduquette.joe.Trace;
 import com.wjduquette.joe.patterns.Pattern;
 import com.wjduquette.joe.scanner.Scanner;
-import com.wjduquette.joe.scanner.SourceBuffer;
-import com.wjduquette.joe.scanner.SourceBuffer.Span;
+import com.wjduquette.joe.SourceBuffer;
+import com.wjduquette.joe.SourceBuffer.Span;
 import com.wjduquette.joe.scanner.Token;
 import com.wjduquette.joe.scanner.TokenType;
