+++ lib/src/main/java/com/wjduquette/joe/clark/Compiler.java
@@ -6,8 +6,8 @@ import com.wjduquette.joe.Trace;
 import com.wjduquette.joe.parser.Expr;
 import com.wjduquette.joe.parser.Parser;
 import com.wjduquette.joe.parser.Stmt;
-import com.wjduquette.joe.scanner.SourceBuffer;
-import com.wjduquette.joe.scanner.SourceBuffer.Span;
+import com.wjduquette.joe.SourceBuffer;
+import com.wjduquette.joe.SourceBuffer.Span;
 import com.wjduquette.joe.scanner.Token;
 import com.wjduquette.joe.scanner.TokenType;
