+++ lib/src/main/java/com/wjduquette/joe/parser/Stmt.java
@@ -1,7 +1,7 @@
 package com.wjduquette.joe.parser;
 
 import com.wjduquette.joe.scanner.Token;
-import com.wjduquette.joe.scanner.SourceBuffer.Span;
+import com.wjduquette.joe.SourceBuffer.Span;
 
 import java.util.List;
