+++ lib/src/main/java/com/wjduquette/joe/scanner/Token.java
@@ -1,6 +1,6 @@
 package com.wjduquette.joe.scanner;
 
-import com.wjduquette.joe.scanner.SourceBuffer.Span;
+import com.wjduquette.joe.SourceBuffer.Span;
 
 /**
  * A token produce by Bert's scanner.
